backend/requirements.txt
fastapi==0.95.1
motor==3.1.2
pydantic==1.10.7
pymongo==4.3.3
python-dotenv==1.0.0
shared==0.0.25
